Fix incomplete Gradle 5.6.4 upgrade (#875)
* Java: Proper upgrade to Gradle 5.6.4

When the Java project was upgraded to Gradle 5.6.4, the Gradle wrapper .jar executable wasn't updated with it. The Gradlew scripts also weren't. This commit updates those.

* Kotlin: Proper upgrade to Gradle 5.6.4

When the Kotlin project was upgraded to Gradle 5.6.4, the Gradle wrapper .jar executable wasn't updated with it. The Gradlew scripts also weren't. This commit updates those.